requirements.txt
# Database
psycopg2-binary==2.9.9
SQLAlchemy==1.4.51

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing and analysis
pandas==2.1.3
numpy==1.26.1

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

setuptools==68.0.0
